[Tests] run `nyc` on all tests; use `tape` runner; add full es-shims test suite
