+++ .gitignore
@@ -5,3 +5,7 @@ node_modules
 npm-shrinkwrap.json
 package-lock.json
 yarn.lock
+
+# coverage
+coverage/
+.nyc_output/
